pyproject.toml
[project]
name = "airflow-intro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow==3.0.0",
    "apache-airflow-providers-postgres==6.1.3",
]
